rename claims.go to user_claims.go
