+++ internal/domain/claims/docs.go
@@ -1,0 +1,2 @@
+// Package claims contains parsed token claims structures.
+package claims
